# api/swagger/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "OpenStorage is a clustered implementation of the Open Storage specification and relies on the OCI runtime.\nIt allows you to run stateful services in containers in a multi-host clustered environment.\nThis document represents the API documentaton of Openstorage, for the GO client please visit:\nhttps://github.com/libopenstorage/openstorage",
    "title": "OSD API.",
    "contact": {
      "url": "https://github.com/libopenstorage/openstorage"
    },
    "license": {
      "name": "APACHE2",
      "url": "https://opensource.org/licenses/Apache-2.0"
    },
    "version": "2.0.0"
  },
  "host": "localhost",
  "basePath": "/v1",
  "paths": {
    "/cluster/alerts/{resource}": {
      "get": {
        "description": "This will return a list of alerts for the requested resource",
        "produces": [
          "application/json"
        ],
        "tags": [
          "cluster"
        ],
        "operationId": "enumerateAlerts",
        "parameters": [
          {
            "type": "integer",
            "description": "Resourcetype to get alerts with.\n0: All\n1: Volume\n2: Node\n3: Cluster\n4: Drive\n",
            "name": "resource",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Alerts object",
            "schema": {
              "$ref": "#/definitions/Alerts"
            }
          }
        }
      }
    },
    "/cluster/alerts/{resource}/{id}": {
      "put": {
        "description": "This will clear alert {id} with resourcetype {resource}",
        "produces": [
          "application/json"
        ],
        "tags": [
          "cluster"
        ],
        "operationId": "clearAlert",
        "parameters": [
          {
            "type": "integer",
            "description": "resourcetype to get alerts with.\n0: All\n1: Volume\n2: Node\n3: Cluster\n4: Drive\n",
            "name": "resource",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "description": "id to get alerts with",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Alerts object",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "delete": {
        "description": "This delete clear alert {id} with resourcetype {resource}",
        "produces": [
          "application/json"
        ],
        "tags": [
          "cluster"
        ],
        "operationId": "deleteAlert",
        "parameters": [
          {
            "type": "integer",
            "description": "resourcetype to get alerts with.\n0: All\n1: Volume\n2: Node\n3: Cluster\n4: Drive\n",
            "name": "resource",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "description": "id to get alerts with",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Alerts object",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/cluster/getnodeidfromip/{idip}": {
      "get": {
        "description": "this will return the node ID for the given node IP",
        "produces": [
          "application/json"
        ],
        "tags": [
          "cluster"
        ],
        "operationId": "GetNodeIdFromIp",
        "parameters": [
          {
            "type": "string",
            "description": "cluster node ip or id",
            "name": "idip",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "cluster node ID",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/cluster/inspect/{id}": {
      "get": {
        "description": "This will return the requested node object",
        "produces": [
          "application/json"
        ],
        "tags": [
          "cluster"
        ],
        "summary": "Inspect cluster Nodes.",
        "operationId": "inspectNode",
        "parameters": [
          {
            "type": "integer",
            "description": "id to get node with",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "a node",
            "schema": {
              "$ref": "#/definitions/Node"
            }
          }
        }
      }
    },
...
    "/cluster/peerstatus": {
      "get": {
        "description": "This will return the peer node status",
        "produces": [
          "application/json"
        ],
        "tags": [
          "node"
        ],
        "operationId": "peerStatus",
        "parameters": [
          {
            "type": "integer",
            "description": "id of the node we want to check.",
            "name": "name",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "node status of requested node",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/cluster/secrets/": {
      "get": {
        "description": "This will return the value/data for given secret key",
        "produces": [
          "application/json"
        ],
        "tags": [
          "secrets"
        ],
        "summary": "Get the seceret value/data for given key",
        "operationId": "getSecret",
        "parameters": [
          {
            "type": "string",
            "description": "secret id/key whose value to be retrived",
            "name": "id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "returns the value/data for given key",
            "schema": {
              "$ref": "#/definitions/GetSecretResponse"
            }
          }
        }
      },
      "put": {
        "description": "This will set secrets data/value against given key",
        "produces": [
          "application/json"
        ],
        "tags": [
          "secrets"
        ],
        "summary": "Set Secret Value",
        "operationId": "setSeceret",
        "parameters": [
          {
            "type": "string",
            "description": "key/id for secrets",
            "name": "id",
            "in": "query",
            "required": true
          },
          {
            "description": "value/data for secrets",
            "name": "secretvalue",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SetSecretRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        }
      }
    },
...
    "/cluster/{id}": {
      "put": {
        "description": "This will shutdown a node (Not Implemented)",
        "produces": [
          "application/json"
        ],
        "tags": [
          "cluster"
        ],
        "operationId": "shutdownNode",
        "parameters": [
          {
            "type": "integer",
            "description": "id to get node with",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "shutdown success",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "delete": {
        "description": "This will delete a node from the cluster",
        "produces": [
          "application/json"
        ],
        "tags": [
          "cluster"
        ],
        "operationId": "deleteNode",
        "parameters": [
          {
            "type": "integer",
            "description": "id to get node with",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "boolean",
            "description": "forceRemove node",
            "name": "forceRemove",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "delete node success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/config/cluster": {
      "get": {
        "description": "This will return the requested cluster configuration object",
        "produces": [
          "application/json"
        ],
        "tags": [
          "config"
        ],
        "summary": "Get cluster configuration.",
        "operationId": "getClusterConfig",
        "responses": {
          "200": {
            "description": "a cluster config",
            "schema": {
              "$ref": "#/definitions/ClusterConfig"
            }
          }
        }
      },
      "post": {
        "description": "This will set the requested cluster configuration",
        "produces": [
          "application/json"
        ],
        "tags": [
          "config"
        ],
        "summary": "Set cluster configuration.",
        "operationId": "setClusterConfig",
        "parameters": [
          {
            "description": "cluster config json",
            "name": "config",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ClusterConfig"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Alerts": {
      "description": "Alerts is an array of Alert objects",
      "type": "object",
      "properties": {
        "alert": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Alert"
          },
          "x-go-name": "Alert"
        }
      },
      "x-go-package": "github.com/libopenstorage/openstorage/api"
    },
...
    "GetSecretResponse": {
      "description": "GetSecretResponse gets secret value for given key",
      "type": "object",
      "properties": {
        "SecretValue": {
          "type": "object"
        }
      },
      "x-go-package": "github.com/libopenstorage/openstorage/secrets"
    },
...
    "Node": {
      "description": "It includes the current physical state (CPU, memory, storage, network usage) as\nwell as the containers running on the system.",
      "type": "object",
      "title": "Node describes the state of a node.",
      "properties": {
        "Avgload": {
          "description": "Average load (percentage)",
          "type": "integer",
          "format": "int64"
        },
        "Cpu": {
          "description": "Cpu usage of the node.",
          "type": "number",
          "format": "double"
        },
        "DataIp": {
          "description": "Data IP",
          "type": "string"
        },
        "Disks": {
          "description": "List of disks on this node.",
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/StorageResource"
          }
        },
        "GenNumber": {
          "description": "GenNumber of the node",
          "type": "integer",
          "format": "uint64"
        },
        "Hostname": {
          "description": "Hostname of this node",
          "type": "string"
        },
        "Id": {
          "description": "Id of the node.",
          "type": "string"
        },
        "MemFree": {
          "description": "Free Memory of the node",
          "type": "integer",
          "format": "uint64"
        },
        "MemTotal": {
          "description": "Total Memory of the node",
          "type": "integer",
          "format": "uint64"
        },
        "MemUsed": {
          "description": "Used Memory of the node",
          "type": "integer",
          "format": "uint64"
        },
        "MgmtIp": {
          "description": "Management IP",
          "type": "string"
        },
        "NodeData": {
          "description": "Node data for this node (EX: Public IP, Provider, City..)",
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "NodeLabels": {
          "description": "User defined labels for node. Key Value pairs",
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "Pools": {
          "description": "List of storage pools this node supports",
          "type": "array",
          "items": {
            "$ref": "#/definitions/StoragePool"
          }
        },
        "StartTime": {
          "description": "Start time of this node",
          "type": "string",
          "format": "date-time"
        },
        "Status": {
          "$ref": "#/definitions/Status"
        },
        "Timestamp": {
          "description": "Timestamp",
          "type": "string",
          "format": "date-time"
        }
      },
      "x-go-package": "github.com/libopenstorage/openstorage/api"
    },
...
    "SetSecretRequest": {
      "description": "SetSecretRequest stores the given value/data against the key",
      "type": "object",
      "properties": {
        "SecretValue": {
          "type": "object"
        }
      },
      "x-go-package": "github.com/libopenstorage/openstorage/secrets"
    },
...
    "Status": {
      "type": "integer",
      "format": "int32",
      "x-go-package": "github.com/libopenstorage/openstorage/api"
    },
...
  },
...
}